move ui test directory
