+++ web/src/tests/rtl/PTeam/AuthAdminCheckbox.test.jsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent, { PointerEventsCheckLevel } from "@testing-library/user-event";
 import React from "react";
 
-import { AuthAdminCheckbox } from "./AuthAdminCheckbox";
+import { AuthAdminCheckbox } from "../../../pages/PTeam/AuthAdminCheckbox";
 
 const mockTemplate = () => {
   throw new Error("Not implemented: You should override mock using jest.fn().");
